LPS-65012 Portal Settings move constant class out of internal and export the package
