+++ modules/apps/foundation/portal-settings/portal-settings-web/src/main/java/com/liferay/portal/settings/web/constants/PortalSettingsPortletKeys.java
@@ -12,17 +12,14 @@
  * details.
  */
 
-package com.liferay.portal.settings.web.internal.constants;
+package com.liferay.portal.settings.web.constants;
 
 /**
- * @author Tomas Polesovsky
+ * @author Philip Jones
  */
-public interface PortalSettingsWebKeys {
+public class PortalSettingsPortletKeys {
 
-	public static final String AUTHENTICATION_DYNAMIC_INCLUDES =
-		"AUTHENTICATION_DYNAMIC_INCLUDES";
-
-	public static final String AUTHENTICATION_TABS_NAMES =
-		"AUTHENTICATION_TABS_NAMES";
+	public static final String PORTAL_SETTINGS =
+		"com_liferay_portal_settings_web_portlet_PortalSettingsPortlet";
 
 }
