+++ modules/apps/foundation/portal-settings/portal-settings-web/src/main/java/com/liferay/portal/settings/web/constants/PortalSettingsWebKeys.java
@@ -12,14 +12,17 @@
  * details.
  */
 
-package com.liferay.portal.settings.web.internal.constants;
+package com.liferay.portal.settings.web.constants;
 
 /**
- * @author Philip Jones
+ * @author Tomas Polesovsky
  */
-public class PortalSettingsPortletKeys {
+public interface PortalSettingsWebKeys {
 
-	public static final String PORTAL_SETTINGS =
-		"com_liferay_portal_settings_web_portlet_PortalSettingsPortlet";
+	public static final String AUTHENTICATION_DYNAMIC_INCLUDES =
+		"AUTHENTICATION_DYNAMIC_INCLUDES";
+
+	public static final String AUTHENTICATION_TABS_NAMES =
+		"AUTHENTICATION_TABS_NAMES";
 
 }
