+++ modules/apps/foundation/portal-settings/portal-settings-web/src/main/resources/com/liferay/portal/settings/web/constants/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
